# docs/release/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "辽轨智能实训能力评估平台",
    "description": "辽宁铁道职业技术学院智能实训能力评估平台 API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/environment/tasks": {
      "post": {
        "tags": [
          "环境检查"
        ],
        "summary": "Create Environment Task",
        "operationId": "create_environment_task_api_v1_environment_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnvironmentCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvironmentTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/environment/tasks/{task_id}": {
      "get": {
        "tags": [
          "环境检查"
        ],
        "summary": "Get Environment Task",
        "operationId": "get_environment_task_api_v1_environment_tasks__task_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvironmentTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/environment/labs/{lab_id}": {
      "get": {
        "tags": [
          "环境检查"
        ],
        "summary": "Get Lab",
        "description": "获取实训室详情",
        "operationId": "get_lab_api_v1_environment_labs__lab_id__get",
        "parameters": [
          {
            "name": "lab_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lab Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/environment/check": {
      "post": {
        "tags": [
          "环境检查"
        ],
        "summary": "Check Environment",
        "description": "上传照片进行环境检查",
        "operationId": "check_environment_api_v1_environment_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnvironmentCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvironmentCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/environment/history/{student_id}": {
      "get": {
        "tags": [
          "环境检查"
        ],
        "summary": "Get Student Check History",
        "description": "学生查看本人结果，教师/管理员查看授权范围结果。",
        "operationId": "get_student_check_history_api_v1_environment_history__student_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Student Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnvironmentCheckResponse"
                  },
                  "title": "Response Get Student Check History Api V1 Environment History  Student Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/environment/checks/{check_id}": {
      "get": {
        "tags": [
          "环境检查"
        ],
        "summary": "Get Check",
        "operationId": "get_check_api_v1_environment_checks__check_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "check_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Check Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvironmentCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/environment/checks/{check_id}/review": {
      "post": {
        "tags": [
          "环境检查"
        ],
        "summary": "Review Check",
        "operationId": "review_check_api_v1_environment_checks__check_id__review_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "check_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Check Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnvironmentReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvironmentCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EnvironmentCheckRequest": {
        "properties": {
          "student_id": {
            "type": "string",
            "title": "Student Id"
          },
          "lab_id": {
            "type": "string",
            "title": "Lab Id"
          },
          "image_base64": {
            "type": "string",
            "title": "Image Base64"
          },
          "score_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Id"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "lab_id",
          "image_base64"
        ],
        "title": "EnvironmentCheckRequest"
      },
      "EnvironmentCheckResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "student_id": {
            "type": "string",
            "title": "Student Id"
          },
          "lab_id": {
            "type": "string",
            "title": "Lab Id"
          },
          "lab_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lab Name"
          },
          "total_score": {
            "type": "integer",
            "title": "Total Score"
          },
          "final_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Score"
          },
          "max_score": {
            "type": "integer",
            "title": "Max Score",
            "default": 100
          },
          "details": {
            "type": "object",
            "title": "Details"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "suggestions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggestions"
          },
          "checked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked At"
          },
          "uploaded_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded Image Url"
          },
          "reference_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Image Url"
          },
          "review_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Status"
          },
          "reviewed_details": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed Details"
          },
          "reviewed_suggestions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reviewed Suggestions"
          },
          "reviewed_suggestions_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed Suggestions Comment"
          },
          "reviewed_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed Summary"
          },
          "review_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Note"
          },
          "reviewer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Name"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "student_id",
          "lab_id",
          "total_score",
          "final_score",
          "details",
          "summary"
        ],
        "title": "EnvironmentCheckResponse"
      },
      "EnvironmentReviewRequest": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "reviewed_details": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed Details"
          },
          "reviewed_suggestions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed Suggestions"
          },
          "reviewed_suggestions_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed Suggestions Comment"
          },
          "reviewed_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed Summary"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "EnvironmentReviewRequest"
      },
      "EnvironmentTaskResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "student_id": {
            "type": "string",
            "title": "Student Id"
          },
          "lab_id": {
            "type": "string",
            "title": "Lab Id"
          },
          "score_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "check_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Check Id"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EnvironmentCheckResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "student_id",
          "lab_id",
          "status"
        ],
        "title": "EnvironmentTaskResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LabResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "building": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Building"
          },
          "floor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Floor"
          },
          "capacity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity"
          },
          "equipment": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Equipment",
            "default": []
          },
          "reference_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Image Url"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "current_students": {
            "type": "integer",
            "title": "Current Students"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "current_students"
        ],
        "title": "LabResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}